Enhance fractal rendering with GMP optimizations and new features: added in-place multiplication and square functions for complex numbers, improved event handling during calculations, and updated GUI to reflect new zoom factor calculations. Adjusted color modes and refined documentation for clarity.

+++ src/SDLGUI.h
@@ -42,5 +42,5 @@ void SDLGUI_MenuBar_Draw (SDL_Surface*, gui*);
 extern void SDLGUI_Button_Draw (SDL_Surface*, gui*, int);
 void SDLGUI_Draw3DBox (SDL_Surface *surface, int x, int y, int w, int h, Uint32 bgcolor, int type);
 void SDLGUI_StateBar_Draw (SDL_Surface* screen, gui* g);
-void SDLGUI_StateBar_Update (SDL_Surface* screen, gui* g, int type, int colorMode, double centerX, double centerY, int zoomFactor, Uint32 renderTime, fractal* f);
+void SDLGUI_StateBar_Update (SDL_Surface* screen, gui* g, int type, int colorMode, double centerX, double centerY, double zoomFactor, Uint32 renderTime, fractal* f);
 void SDLGUI_StateBar_Progress (SDL_Surface* screen, gui* g, int percent, const char* task);
